requirements/development.txt
pre-commit==2.17.0
black==21.12b0

pytest==6.2.5
pytest-django==4.5.2
selenium==4.1.0

factory-boy==3.2.1
